Change database traces from info -> debug (#135)
The database insertion logging is very verbose, and makes it hard to see
what the application is doing (particularly when syncing / catching up).
This changes those from `info!` to `debug!` so we can have a cleaner log
by default while still being able to access those insertion logs if
desired.